requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dateutil>=2.8.2
jinja2>=3.1.0
pytest>=7.4.0
pytest-mock>=3.12.0
freezegun>=1.2.0
anthropic>=0.39.0
pymupdf>=1.24.0
flask>=3.0.0
gunicorn>=22.0.0
